requirements.txt
autopep8==1.6.0
greenlet==1.1.2
numpy==1.22.4
pandas==1.4.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.37
toml==0.10.2
